requirements.txt
fastapi
uvicorn
python-dotenv
firebase-admin
httpx
passlib
pyjwt
email-validator
cryptography

